feat(settings): OutgoingEmailEditor + combined /settings/outgoing (#234)
Slice 8 of the Settings redesign.

OutgoingEmailEditor deep module: `<OutgoingEmailEditor kind="invoice" |
"contract" | "payment-link" />`. The shell owns load / dirty tracking /
PATCH save uniformly; per-kind config picks the endpoint, title, and form
body. The three previous page bodies collapse into kind-specific
renderForm functions and the payment template field moves to
components/settings so the redirected route directory holds only the
redirect.

Combined section at /settings/outgoing wires three tabs (Invoices,
Contracts, Payment links) via SettingsTabs, with Invoices as the default.

The misleadingly-named "Contracts" and "Outgoing Emails" sidebar entries
collapse into a single "Outgoing Emails" entry pointing at the new
combined route. The hidden /settings/invoices page is surfaced as the
Invoices tab.

Redirects:
- /settings/invoices  -> /settings/outgoing?tab=invoices
- /settings/contracts -> /settings/outgoing?tab=contracts
- /settings/payments  -> /settings/outgoing?tab=payment-links

Tests: load / dirty / save endpoint+payload for each kind (9 tests), plus
a smoke test for the /settings/outgoing shell.

Closes #234.
