+++ src/lib/settings-nav.ts
@@ -39,11 +39,10 @@ export const settingsNavItems: SettingsNavItem[] = [
   { href: "/settings/email", label: "Email Accounts", icon: Mail },
   { href: "/settings/signatures", label: "Email Signatures", icon: FileSignature },
   { href: "/settings/contract-templates", label: "Contract Templates", icon: FileText },
-  { href: "/settings/contracts", label: "Contracts", icon: Send },
+  { href: "/settings/outgoing", label: "Outgoing Emails", icon: Send },
   { href: "/settings/notifications", label: "Notifications", icon: Bell },
   { href: "/settings/accounting", label: "Accounting", icon: Link2 },
   { href: "/settings/stripe", label: "Stripe Payments", icon: CreditCard },
-  { href: "/settings/payments", label: "Outgoing Emails", icon: Mail },
   { href: "/settings/reports", label: "Reports", icon: FileText },
   { href: "/settings/export", label: "Data Export", icon: Download },
   { href: "/settings/knowledge", label: "Knowledge Base", icon: BookOpen },
